more fixes (cast!)
git-svn-id: svn+ssh://s4.hekkelman.net/home/maarten/svnroot/salt/trunk/libassh@179 49180b5e-4fd1-4af1-8649-882265d54e9a

+++ assh/packet.hpp
@@ -84,7 +84,7 @@ class opacket
 	// for ranges:
 	opacket&		operator<<(const std::pair<const char*,std::size_t>& v)
 					{
-						operator<<(v.second);
+						operator<<(uint32(v.second));
 						m_data.insert(m_data.end(), reinterpret_cast<const uint8*>(v.first),
 							reinterpret_cast<const uint8*>(v.first + v.second));
 						return *this;
